+++ test/libsolidity/syntaxTests/controlFlow/storageReturn/assembly/switch_fine.sol
@@ -1,20 +1,14 @@
 contract C {
     struct S { bool f; }
     S s;
-    function f(uint256 a) internal pure returns (S storage c) {
-        assembly {
-            switch a
-            default { c_slot := s_slot }
-        }
-    }
-    function g(bool flag) internal pure returns (S storage c) {
+    function f(bool flag) internal pure returns (S storage c) {
         assembly {
             switch flag
             case 0 { c_slot := s_slot }
             default { c_slot := s_slot }
         }
     }
-    function h(uint256 a) internal pure returns (S storage c) {
+    function g(uint256 a) internal pure returns (S storage c) {
         assembly {
             switch a
             case 0 { revert(0, 0) }
@@ -23,4 +17,3 @@ contract C {
     }
 }
 // ----
-// Warning: (142-191): "switch" statement with only a default case.