+++ Changelog.md
@@ -4,6 +4,7 @@ Language Features:
 
 
 Compiler Features:
+ * Yul: Raise warning for switch statements that only have a default and no other cases.
 
 
 Bugfixes: